+++ src/System.Windows.Forms.Design/src/Resources/SR.resx
@@ -1324,19 +1324,19 @@ Press Ctrl+Enter to accept Text.</value>
     <value>Horizontal Splitter Orientation.</value>
   </data>
   <data name="ImageListActionList_ChooseImagesDescription" xml:space="preserve">
-    <value />
+    <value>Opens the Images collection editor</value>
   </data>
   <data name="ImageListActionList_ChooseImagesDisplayName" xml:space="preserve">
     <value>Choose images</value>
   </data>
   <data name="ImageListActionList_ColorDepthDescription" xml:space="preserve">
-    <value />
+    <value>Changes the bit depth images imported into this ImageList are stored at</value>
   </data>
   <data name="ImageListActionList_ColorDepthDisplayName" xml:space="preserve">
     <value>Image Bit Depth</value>
   </data>
   <data name="ImageListActionList_ImageSizeDescription" xml:space="preserve">
-    <value />
+    <value>Changes the size images imported into this ImageList are stored at</value>
   </data>
   <data name="ImageListActionList_ImageSizeDisplayName" xml:space="preserve">
     <value>Image Size</value>
